Extract all & sizes are now aware of the selected bundles
Fixed total size using uint which made it so 2.8gb is the limit individual files are still uint since there aren't files that big afaik.
Removed datacontext from entry (use utils CurrentWindow instead)

+++ Services/Utils.cs
@@ -81,7 +81,7 @@ namespace DieselBundleViewer.Services
                 return dir + "/" + dir2;
         }
 
-        public static string FriendlySize(uint size)
+        public static string FriendlySize(ulong size)
         {
             if (size < 1024)
                 return size.ToString() + " B";
